Moved csvy parser to model reader
